# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Nave - Agente de IA","description":"API para gerenciamento de usuários, consentimentos e mensagens do agente de IA Nave.","version":"1.0.0"},"servers":[{"url":"http://localhost:8000","description":"Servidor local para desenvolvimento"}],"paths":{"/auth/":{"get":{"tags":["auth"],"summary":"Autenticar","operationId":"autenticar_auth__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/criar_usuario":{"post":{"tags":["auth"],"summary":"Criar Usuario","operationId":"criar_usuario_auth_criar_usuario_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuarioschema"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoguinSchema"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh Token","operationId":"refresh_token_auth_refresh_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/consents/":{"get":{"tags":["consents"],"summary":"Listar Consents","description":"Lista todas as permissões do agente para o usuário autenticado.","operationId":"listar_consents_consents__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ConsentResponseSchema"},"type":"array","title":"Response Listar Consents Consents  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["consents"],"summary":"Criar Consent","description":"Concede uma nova permissão ao agente de IA.","operationId":"criar_consent_consents__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentCreateSchema"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/consents/{permissao}":{"delete":{"tags":["consents"],"summary":"Revogar Consent","description":"Revoga (desativa) uma permissão do agente. Não apaga o registro.","operationId":"revogar_consent_consents__permissao__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"permissao","in":"path","required":true,"schema":{"type":"string","title":"Permissao"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/":{"get":{"tags":["message"],"summary":"Listar Historico","description":"Retorna todo o histórico de mensagens do usuário autenticado.","operationId":"listar_historico_message__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MessageResponseSchema"},"type":"array","title":"Response Listar Historico Message  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["message"],"summary":"Enviar Mensagem","description":"Recebe a mensagem do usuário, chama o agente de IA com o histórico\ncompleto e salva tanto a mensagem do usuário quanto a resposta.","operationId":"enviar_mensagem_message__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreateSchema"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["message"],"summary":"Limpar Historico","description":"Apaga todo o histórico de mensagens do usuário autenticado.","operationId":"limpar_historico_message__delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"ConsentCreateSchema":{"properties":{"permissao":{"type":"string","title":"Permissao"}},"type":"object","required":["permissao"],"title":"ConsentCreateSchema"},"ConsentResponseSchema":{"properties":{"permissao":{"type":"string","title":"Permissao"},"id":{"type":"integer","title":"Id"},"usuario_id":{"type":"integer","title":"Usuario Id"},"ativo":{"type":"boolean","title":"Ativo"},"criado_em":{"type":"string","format":"date-time","title":"Criado Em"}},"type":"object","required":["permissao","id","usuario_id","ativo","criado_em"],"title":"ConsentResponseSchema"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoguinSchema":{"properties":{"email":{"type":"string","title":"Email"},"senha":{"type":"string","title":"Senha"}},"type":"object","required":["email","senha"],"title":"LoguinSchema"},"MessageCreateSchema":{"properties":{"conteudo":{"type":"string","title":"Conteudo"}},"type":"object","required":["conteudo"],"title":"MessageCreateSchema"},"MessageResponseSchema":{"properties":{"id":{"type":"integer","title":"Id"},"usuario_id":{"type":"integer","title":"Usuario Id"},"role":{"type":"string","title":"Role"},"conteudo":{"type":"string","title":"Conteudo"},"criado_em":{"type":"string","format":"date-time","title":"Criado Em"}},"type":"object","required":["id","usuario_id","role","conteudo","criado_em"],"title":"MessageResponseSchema"},"Usuarioschema":{"properties":{"nome":{"type":"string","title":"Nome"},"email":{"type":"string","title":"Email"},"senha":{"type":"string","title":"Senha"},"ativo":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Ativo"},"admin":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Admin"}},"type":"object","required":["nome","email","senha","ativo","admin"],"title":"Usuarioschema"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/login"}}}}}}